backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
gunicorn==21.2.0
